requirements.txt
# Dependencies for Discord Status Bot
discord.py-self==2.0.1
rich==13.7.0
flask==2.3.3
